# Lazily initialize the time zone cache