--- a/src/types/timezone.jl
+++ b/src/types/timezone.jl
@@ -2,14 +2,31 @@
 # to this structure can result in inconsistent behaviour.
 # Do not access this object directly, instead use `get_tz_cache()` to access the cache.
 const _TZ_CACHE = Dict{String,Tuple{TimeZone,Class}}()
+const _TZ_CACHE_LOCK = ReentrantLock()
+const _TZ_CACHE_INITIALIZED = Threads.Atomic{Bool}(false)
 
-function _reload_cache(compiled_dir::AbstractString)
-    _reload_cache!(_TZ_CACHE, compiled_dir)
-    !isempty(_TZ_CACHE) || error("Cache remains empty after loading")
-    return nothing
+function _init_tz_cache()
+    # Write out our compiled tzdata representations into a scratchspace
+    desired_version = TZData.tzdata_version()
+
+    _COMPILED_DIR[] = if desired_version == TZJData.TZDATA_VERSION
+        TZJData.ARTIFACT_DIR
+    else
+        # @info "Loading tzdata $desired_version"
+        TZData.build(desired_version, _scratch_dir())
+    end
+
+    # Load the pre-computed TZData into memory.
+    return _reload_tz_cache(_COMPILED_DIR[])
 end
 
-function _reload_cache!(cache::AbstractDict, compiled_dir::AbstractString)
+function _reload_tz_cache(compiled_dir::AbstractString)
+    _reload_tz_cache!(_TZ_CACHE, compiled_dir)
+    !isempty(_TZ_CACHE) || error("Cache remains empty after loading")
+    return _TZ_CACHE
+end
+
+function _reload_tz_cache!(cache::AbstractDict, compiled_dir::AbstractString)
     empty!(cache)
     check = Tuple{String,String}[(compiled_dir, "")]
 
@@ -31,37 +48,17 @@ function _reload_cache!(cache::AbstractDict, compiled_dir::AbstractString)
     return cache
 end
 
-const _TZ_CACHE_INITIALIZED = Threads.Atomic{Bool}(false)
-function initialize_tz_cache()
-    # Write out our compiled tzdata representations into a scratchspace
-    desired_version = TZData.tzdata_version()
-
-    _COMPILED_DIR[] = if desired_version == TZJData.TZDATA_VERSION
-        TZJData.ARTIFACT_DIR
-    else
-        # @info "Loading tzdata $desired_version"
-        TZData.build(desired_version, _scratch_dir())
-    end
-
-    # Load the pre-computed TZData into memory.
-    _reload_cache(_COMPILED_DIR[])
-    _TZ_CACHE_INITIALIZED[] = true
-
-    return nothing
-end
-
-const _TZ_CACHE_LOCK = ReentrantLock()
-function get_tz_cache()
-    if  !_TZ_CACHE_INITIALIZED[]
+function _get_tz_cache()
+    if !_TZ_CACHE_INITIALIZED[]
         lock(_TZ_CACHE_LOCK) do
             if !_TZ_CACHE_INITIALIZED[]
-                initialize_tz_cache()
+                _init_tz_cache()
+                _TZ_CACHE_INITIALIZED[] = true
             end
         end
     end
     return _TZ_CACHE
 end
-
 
 """
     TimeZone(str::AbstractString) -> TimeZone
@@ -104,7 +101,7 @@ US/Pacific (UTC-8/UTC-7)
 TimeZone(::AbstractString, ::Class)
 
 function TimeZone(str::AbstractString, mask::Class=Class(:DEFAULT))
-    tz, class = get(get_tz_cache(), str) do
+    tz, class = get(_get_tz_cache(), str) do
         if occursin(FIXED_TIME_ZONE_REGEX, str)
             FixedTimeZone(str), Class(:FIXED)
         else
@@ -149,6 +146,6 @@ function istimezone(str::AbstractString, mask::Class=Class(:DEFAULT))
     end
 
     # Checks against pre-compiled time zones
-    class = get(get_tz_cache(), str, (UTC_ZERO, Class(:NONE)))[2]
+    class = get(_get_tz_cache(), str, (UTC_ZERO, Class(:NONE)))[2]
     return mask & class != Class(:NONE)
 end