--- a/src/types/timezone.jl
+++ b/src/types/timezone.jl
@@ -12,7 +12,6 @@ function _init_tz_cache()
     _COMPILED_DIR[] = if desired_version == TZJData.TZDATA_VERSION
         TZJData.ARTIFACT_DIR
     else
-        # @info "Loading tzdata $desired_version"
         TZData.build(desired_version, _scratch_dir())
     end