This avoids having to compile a bunch of code in `__init__` and improves the load time of the package (on Julia 1.11) from:

```julia
julia> @time_imports import TimeZones
...
              ┌ 2.6 ms TimeZones.TZData.__init__() 92.83% compilation time
               ├ 237.9 ms TimeZones.__init__() 94.11% compilation time (89% recompilation)
    276.3 ms  TimeZones 81.90% compilation time (88% recompilation)
```

```julia
julia> @time_imports import TimeZones
...
     17.5 ms  TimeZones 16.03% compilation time
```

The change in performance of the functions is (in my opinion) acceptable.

After:

```julia
julia> using BenchmarkTools, TimeZones

julia> @btime istimezone("Europe/Warsaw")
  48.068 ns (1 allocation: 16 bytes)
true
```

Before:

```julia
julia> @btime istimezone("Europe/Warsaw")
  45.356 ns (1 allocation: 16 bytes)
true
```