@omus 2024-05-17
Posting some benchmarks from my machine so I can do some refactoring without losing the performance gains.

Before this PR with Julia 1.11.0-beta1 on an M1 MacBook
```
julia> @time_imports import TimeZones
...
               ┌ 4.4 ms TimeZones.TZData.__init__() 53.17% compilation time
               ├ 318.6 ms TimeZones.__init__() 85.10% compilation time (93% recompilation)
    690.8 ms  TimeZones 59.42% compilation time (85% recompilation)

julia> using BenchmarkTools, TimeZones

julia> @btime istimezone("Europe/Warsaw");
  77.314 ns (1 allocation: 16 bytes)
```

With this PR with Julia 1.11.0-beta1 on an M1 MacBook
```
julia> @time_imports import TimeZones
...
               ┌ 3.7 ms TimeZones.TZData.__init__() 56.80% compilation time
               ├ 0.0 ms TimeZones.__init__()
    107.0 ms  TimeZones 83.43% compilation time (53% recompilation)

julia> using BenchmarkTools, TimeZones

julia> @btime istimezone("Europe/Warsaw");
  80.966 ns (1 allocation: 16 bytes)
```