@omus 2024-05-17
<details>
<summary>Iterating on the `istimezone` internals</summary>

### Iteration 1

```julia
function istimezone(str::AbstractString, mask::Class=Class(:DEFAULT))
    # Start by performing quick FIXED class test
    if mask & Class(:FIXED) != Class(:NONE) && occursin(FIXED_TIME_ZONE_REGEX, str)
        return true
    end

    # Checks against pre-compiled time zones
    tz, class = get(get_tz_cache(), str, (nothing, Class(:NONE)))
    return tz !== nothing && mask & class != Class(:NONE)
end
```

```julia
julia> @time_imports import TimeZones
...
               ┌ 2.7 ms TimeZones.TZData.__init__() 77.28% compilation time
               ├ 0.0 ms TimeZones.__init__()
    106.6 ms  TimeZones 84.49% compilation time (53% recompilation)

julia> using BenchmarkTools, TimeZones

julia> @btime istimezone("Europe/Warsaw");
  95.032 ns (2 allocations: 64 bytes)
```

### Iteration 2

```julia
function istimezone(str::AbstractString, mask::Class=Class(:DEFAULT))
    # Start by performing quick FIXED class test
    if mask & Class(:FIXED) != Class(:NONE) && occursin(FIXED_TIME_ZONE_REGEX, str)
        return true
    end

    # Checks against pre-compiled time zones
    tz_class = get(get_tz_cache(), str, nothing)
    tz_class === nothing && return false
    tz, class = tz_class
    return mask & class != Class(:NONE)
end
```


```julia
julia> @time_imports import TimeZones
...
               ┌ 2.5 ms TimeZones.TZData.__init__() 78.26% compilation time
               ├ 0.0 ms TimeZones.__init__()
    103.5 ms  TimeZones 83.75% compilation time (52% recompilation)

julia> using BenchmarkTools, TimeZones

julia> @btime istimezone("Europe/Warsaw");
  75.789 ns (1 allocation: 16 bytes)
```

### Iteration 3

```julia
function istimezone(str::AbstractString, mask::Class=Class(:DEFAULT))
    # Start by performing quick FIXED class test
    if mask & Class(:FIXED) != Class(:NONE) && occursin(FIXED_TIME_ZONE_REGEX, str)
        return true
    end

    # Checks against pre-compiled time zones
    _, class = get(get_tz_cache(), str, (UTC_ZERO, Class(:NONE)))
    return mask & class != Class(:NONE)
end
```

```julia
julia> @time_imports import TimeZones
...
               ┌ 3.1 ms TimeZones.TZData.__init__() 66.09% compilation time
               ├ 0.0 ms TimeZones.__init__()
    103.3 ms  TimeZones 83.89% compilation time (52% recompilation)

julia> using BenchmarkTools, TimeZones

julia> @btime istimezone("Europe/Warsaw");
  80.492 ns (1 allocation: 16 bytes)
```

### Iteration 4

```julia
function istimezone(str::AbstractString, mask::Class=Class(:DEFAULT))
    # Start by performing quick FIXED class test
    if mask & Class(:FIXED) != Class(:NONE) && occursin(FIXED_TIME_ZONE_REGEX, str)
        return true
    end

    # Checks against pre-compiled time zones
    class = get(get_tz_cache(), str, (UTC_ZERO, Class(:NONE)))[2]
    return mask & class != Class(:NONE)
end
```

```julia
julia> @time_imports import TimeZones
...
               ┌ 2.8 ms TimeZones.TZData.__init__() 79.89% compilation time
               ├ 0.0 ms TimeZones.__init__()
    102.1 ms  TimeZones 83.99% compilation time (54% recompilation)

julia> using BenchmarkTools, TimeZones

julia> @btime istimezone("Europe/Warsaw");
  75.610 ns (1 allocation: 16 bytes)
```

</details>