@omus 2024-05-17
Commit 5701f3a3c67c7dd3da9579c22414e9e7236a89b6

```julia
julia> @time_imports import TimeZones
...
               ┌ 3.1 ms TimeZones.TZData.__init__() 77.56% compilation time
               ├ 0.0 ms TimeZones.__init__()
    106.5 ms  TimeZones 84.26% compilation time (54% recompilation)

julia> using BenchmarkTools, TimeZones

julia> @btime istimezone("Europe/Warsaw");
  78.689 ns (1 allocation: 16 bytes)
```

Commit b5948fc5e593d2a52087d5b835e1dcd05876f31d

```julia
julia> @time_imports import TimeZones
...
      0.6 ms  Mocking
               ┌ 2.5 ms TimeZones.TZData.__init__() 81.56% compilation time
               ├ 0.0 ms TimeZones.__init__()
    108.0 ms  TimeZones 79.75% compilation time (53% recompilation)

julia> using BenchmarkTools, TimeZones

julia> @btime istimezone("Europe/Warsaw");
  75.703 ns (1 allocation: 16 bytes)
```

Looks like I broke some tests in the process