@omus 2024-05-17
Close to the finish line on this one. I'd like to recheck the performance with the fixed tests yet.